Design of Configurations

A lightweight wrapper around a ParamSet and a data.table::data.table(), where the latter is a design of configurations produced from the former - e.g., by calling a generate_design_grid() or by sampling.

Public fields

param_set

(ParamSet).

data

(data.table::data.table())
Stored data.

Methods

Public methods

Design$new()

Design$format()

Design$print()

Design$transpose()

Design$clone()

Method new()

Creates a new instance of this R6 class.

Usage

Design$new(param_set, data, remove_dupl)

Arguments

param_set

(ParamSet).

data

(data.table::data.table())
Stored data.

remove_dupl

(logical(1))
Remove duplicates?

Method format()

Helper for print outputs.

Usage

Design$format()

Method print()

Printer.

Usage

Design$print(...)

Arguments

...

(ignored).

Method transpose()

Converts data into a list of lists of row-configurations, possibly removes NA entries of inactive parameter values due to unsatisfied dependencies, and possibly calls the trafo function of the ParamSet.

Usage

Design$transpose(filter_na = TRUE, trafo = TRUE)

Arguments

filter_na

(logical(1))
Should NA entries of inactive parameter values due to unsatisfied dependencies be removed?

trafo

(logical(1))
Should the trafo function of the ParamSet be called?

Method clone()

The objects of this class are cloneable with this method.

Usage

Design$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.