Base Class for ParamSetBase

A R6Class to represent set of parameters.

ParamSetBase

Format

R6Class object

Value

[ParamSetBase].

Member Variables

params

[list]
List of the Params

trafo

[function(x, dict, tags)]
A function that returns a list of transformed x values. Has to work vectorized and also return untransformed x values. The function takes a list x of all parameter values, additionally the dictionary linked to the ParamSet. tags is a named list that contains the tags for each Param in x.

restriction

[quote]
A quoted expression (quote()) that is evaluated on all parameter values to check if they are feasible. It has to be evaluated to TRUE so that the parameter value is valid. The expression has to work on vectors of values.

Inherited from ParamNode:

id

[character]
Unique id.

storage_type

[character(1)]
How are values for this Parameter represented as an R object.

check

[function]
A function that returns TRUE for valid parameter values and a character if the value is not valid.

assert

[function]
A assert_ion generated by makeAssertionFunction from the check function.

test

[function]
A test generated by makeTestFunction from the check function.

tags

[character]
A vector of characters that indicate special properties of this Parameter.

Methods

none

Inherited from ParamNode:

sample(n)

[function]
Function to sample Parameter values. Always returns a data.table.

denorm(x)

[function]
Takes a named list. Each list item contains a numerical vector with values between [0,1]. These values will be mapped to values of the Parameter Space.

Active Bindings

dictionary

[list|environment]
A dictionary that additional values that might be important for the transformation function, like n for the number of observations.

Inherited from ParamNode

has_finite_bounds

logical(1)
Does the Parameter have finite bounds.