Untyped Parameter Object

A R6Class to represent untyped parameters.

ParamUntyped

Format

R6Class object

Value

[ParamUntyped].

Member Variables

none

Inherited from ParamSimple:

default

[any]
default value.

special_vals

[any]
Special values this parameter is allowed to take that are within the defined space.

Inherited from ParamNode:

id

[character]
Unique id.

storage_type

[character(1)]
How are values for this Parameter represented as an R object.

check

[function]
A function that returns TRUE for valid parameter values and a character if the value is not valid.

assert

[function]
A assert_ion generated by makeAssertionFunction from the check function.

test

[function]
A test generated by makeTestFunction from the check function.

tags

[character]
A vector of characters that indicate special properties of this Parameter.

Methods

none

Inherited from ParamSimple

sampleVector(n)

[function]
samples n Parameter Values.

denorm_vector(x)

[function]
Takes a vector with values between [0,1] and maps them to values of the Parameter.

Inherited from ParamNode:

sample(n)

[function]
Function to sample Parameter values. Always returns a data.table.

denorm(x)

[function]
Takes a named list. Each list item contains a numerical vector with values between [0,1]. These values will be mapped to values of the Parameter Space.

Active Bindings

none

Inherited from ParamSimple

none

Inherited from ParamNode:

has_finite_bounds

logical(1)
Does the Parameter have finite bounds.

See also