Sampler
Source: R/Sampler.R

Random sampler for an arbitrary ParamSet.

Sampler

Format

R6Class object. Abstract base class.

Public members / active bindings

param_set :: ParamSet
Domain / support of the distribution we want to sample from.

Public methods

new(param_set)
ParamSet -> self
Abstract, only inheriting subclasses call this. Param set is cloned on construction.

sample(n)
integer(1) -> Design
Sample n values from the distribution.

Private methods / Internals

.sample(n)
integer(1) -> data.table
Inheriting sublcasses have to implement thus, called from sample()

See also