Sampler1DNormal Class

Normal sampling (potentially truncated) for ParamDbl.

See also

Super classes

paradox::Sampler -> paradox::Sampler1D -> paradox::Sampler1DRfun -> Sampler1DNormal

Active bindings

mean

(numeric(1))
Mean parameter of the normal distribution.

sd

(numeric(1))
SD parameter of the normal distribution.

Methods

Public methods

Sampler1DNormal$new()

Sampler1DNormal$clone()

Inherited methods

paradox::Sampler$format()

paradox::Sampler$print()

paradox::Sampler$sample()

Method new()

Creates a new instance of this R6 class.

Usage

Sampler1DNormal$new(param, mean = NULL, sd = NULL)

Arguments

param

(Param)
Domain / support of the distribution we want to sample from.

mean

(numeric(1))
Mean parameter of the normal distribution. Default is mean(c(param$lower, param$upper).

sd

(numeric(1))
SD parameter of the normal distribution. Default is (param$upper - param$lower)/4.

Method clone()

The objects of this class are cloneable with this method.

Usage

Sampler1DNormal$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.