SamplerUnif Class

Uniform random sampling for an arbitrary (bounded) ParamSet. Constructs 1 uniform sampler per Param, then passes them to SamplerHierarchical. Hence, also works for ParamSets sets with dependencies.

See also

Super classes

paradox::Sampler -> paradox::SamplerHierarchical -> SamplerUnif

Methods

Public methods

SamplerUnif$new()

SamplerUnif$clone()

Inherited methods

paradox::Sampler$format()

paradox::Sampler$print()

paradox::Sampler$sample()

Method new()

Creates a new instance of this R6 class.

Usage

SamplerUnif$new(param_set)

Arguments

param_set

(ParamSet)
Domain / support of the distribution we want to sample from. ParamSet is cloned on construction.

Method clone()

The objects of this class are cloneable with this method.

Usage

SamplerUnif$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.