Param Dependency

A dependency object for a parameter, connects a subordinate parameter with its superordinate parent, and stores a condition on the parents value. In simpler words: we can encode stuff like "foo is only valid, if bar='a'". Constructor is mainly internally called, regular user-entry point is add_dep of ParamSet, the param set maintains an internal list of Dependency objects.

Dependency

Format

An object of class R6ClassGenerator of length 24.

Public members / active bindings

param :: Param The dependent param.

parent :: Param The (categorical) param this param depends on.

cond :: Condition Condition of the dependency.

Public methods

new(param, parent, cond)
Param, Param Condition -> self