Design of Configurations

A lightweight wrapper around a ParamSet and a data.table::data.table(), where the latter is a design of configurations produced from the former - e.g., by calling a generate_design_grid() or by sampling.

Format

R6::R6Class object.

Construction

c = Design$new(param_set, data, remove_dupl)

Note that the first 2 arguments are NOT cloned during construction!

param_set :: ParamSet.

data :: data.table::data.table()
Right hand side of the condition.

remove_dupl :: logical(1)
Remove duplicates?

Fields

param_set :: ParamSet.

data :: data.table::data.table()
Stored data.

Methods

transpose(filter_na = TRUE, trafo = TRUE)
(logical(1), logical(1)) -> list() of list()
Converts data into a list of lists of row-configurations, possibly removes NA entries of inactive parameter values due to unsatisfied dependencies, and possibly calls the trafo function of the ParamSet.