Param Object
Source: R/Param.R

Abstract base class for params and inheriting concrete param subclasses.

Tags: Currently, tags can be used by users in a very custom manner, but one tag is specifically handled: 'required' implied that the parameters has to be given when setting values in ParamSet.

Param

Format

R6Class object.

Public members / active bindings

id :: character(1)
ID of this param.

class :: character(1)
Param R6 class name. Read-only.

lower :: numeric(1)
Lower bound for dbl/int params, can be -Inf. NA if param is not a number.

upper :: numeric(1)
Upper bound for dbl/int params, can be +Inf. NA if param is not a number.

levels :: character | logical | NULL
Allowed levels for categorical params, NULL if param is not categorical.

nlevels :: numeric(1)
Number of categorical levels per parameter, Inf for unbounded ints or any dbl. Read-only.

is_bounded :: logical(1)
Does param have a finitely bounded domain? Read-only.

special_vals :: list
Arbitrary special values this parameter is allowed to take, to make it feasible. This allows extending the domain of the param. This is only used in feasibility checks, neither in generating designs nor sampling.

default :: any
Default value. Can be from param domain or special_vals. Has value NO_DEF if no default is there - NULL could be a valid default.

has_default :: logical(1)
Is a default there?

storage_type :: character(1)
Data type when values of this param are stored in a data table or sampled. Read-only.

tags :: character
Can be used to group and subset params.

is_number :: logical(1)
TRUE iff Param is dbl or int.

is_categ :: logical(1)
TRUE iff Param is fct or lgl.

Public methods

Param$new(id, special_vals = list(), default = NO_DEF, tags = character(0L))
character(1), list, any, character -> self
Constructor of abstract base class, only called by inheriting classes. See meaning of id, special_vals, default, tags in member section.

Param$Dbl$new(id, lower, upper, special_vals, default, tags)
character(1), numeric(1), numeric(1), list, any, character -> self
Constructor for double-scalar-params. Box-constraint bounds can be set, or be Inf.

Param$Int$new(id, lower, upper, special_vals, default, tags)
character(1), numeric(1), numeric(1), list, any, character -> self
Constructor for int-scalar-params. Box-constraint bounds can be set, or be Inf; lower is set to its integer ceiling and 'upper' to its integer floor value.

ParamFct$new(id, values, special_vals, default, tags)
character(1), character, list, any, character -> self
Constructor for categorical/factor-like params; slight misnomer as it accepts only strings, from its defined set of categorical values.

Param$Lgl$new(id, special_vals, default, tags)
character(1), list, any, character -> self
Constructor for logical-scalar-params.

ParamUty$new(id, default, tags, custom_check)
character(1), any, character, function(x) -> self
Untyped parameters, can be used to bypass any complicated feasibility checks, when a param is of truly complex type, as checks for this param are always feasible. OTOH we cannot perform meaningful operations like sampling or generating designs with this param. User can pass a custom_check to specify feasibility checks for custom types.

test(x), check(x), assert(x)
Three checkmate-like check-functions. Take a value from the domain of the param, and check if it is feasible. A value is feasible if it is of the same storage_type, inside of the bounds or from special_vals.

qunif(x)
numeric(n) -> vector(n)
Takes values from [0,1] and map them, regularly distributed, to the domain of the param. Think of: quantile function or the usecse to map a uniform-[0,1] random variable into a uniform sample from this param.

rep(n)
integer(1) -> ParamSet
Repeats this param n-times (by cloning); each param is named "<id>rep<k>" and gets additional tag "<id>_rep".

Further public methods for ParamDbl, ParamInt

range :: numeric(2)
Lower and upper bound as 2-dim-vector.

span :: numeric(1)
Difference of upper - lower.

S3 methods and type converters

as.data.table()
Converts param to datatable with 1 row. See ParamSet.