ParamSet

A set of Param objects. Please note that when creating a set or adding to it, the params of the resulting set have to be uniquely named with IDs with valid R names.

ParamSet

Format

An object of class R6ClassGenerator of length 24.

Public members / active bindings

set_id :: character(1) ID of this param set. Settable.

params :: named list of Param Contained parameters, named with their respective IDs. NB: The returned list contains references, so you can potentially change the objects of the param set by writing to them.

length :: integer(1) Number of contained params. Read-only.

is_empty :: logical(1) Is the param set empty? Read-only.

class :: named character Param classes of contained parameters. Named with param IDs. Read-only.

lower :: named double Lower bounds of parameters, NA if param is not a number. Named with param IDs. Read-only.

upper :: named double Upper bounds of parameters, NA if param is not a number. Named with param IDs. Read-only.

values :: named list List of character vectors of allowed categorical values of contained parameters, NULL if param is not categorical. Named with param IDs. Read-only.

nlevels :: named double Number of categorical levels per parameter, Inf for unbounded ints or any dbl. Named with param IDs. Read-only.

is_bounded :: named logical(1) Do all parameters have finite bounds? Named with param IDs. Read-only.

special_vals :: named list of list
Special values for all parameters. Named with param IDs. Read-only.

storage_type :: character
Data types of params when stored in tables. Named with param IDs. Read-only.

tags :: named list of character
Can be used to group and subset params. Named with param IDs. Read-only.

defaults :: named list
Default values of all params. If no default exists, element is not present. Named with param IDs. Read-only.

is_number :: named logical Position is TRUE iff Param is dbl or int. Named with param IDs. Read-only.

is_categ :: named logical Position is TRUE iff Param is fct or lgl. Named with param IDs. Read-only.

trafo :: function(x, param_set) -> named list
Transformation function. Settable. User has to pass a function(x, param_set), of the form named list, ParamSet -> named list. The function is responsible to transform a feasible configuration into another encoding, before potentially evaluating the configuration with the target algorithm. For the output, not many things have to hold. It needs to have unique names, and the target algorithm has to accept the configuration. For convenience, the self-paramset is also passed in, if you need some info from it (e.g. tags). Is NULL by default, and you can set it to NULL to switch the transformation off.

has_trafo :: logical(1)
Has the set a trafo` function?

deps :: list of Dependency
Parameter dependency objects, each element of the list is internally created by a call to add_dep.

has_deps :: logical(1)
Has the set param dependencies?

deps_on :: data.table
Table has cols id (character(1)) and dep_parent (list of character) and dep (list of Dependency). List all (direct) dependency parents of a param, through parameter IDs. Table has one row per param and is in the same order as ids(). The last col lists all dependencies that the row-ID has.

Public methods

new(params)
list of Param -> self Deep-clones all passed param objects.

ids(class = NULL, is_bounded = NULL, tags = NULL)
character, logical(1), character -> character Retrieves IDs of contained params based on some selections, NULL means no restriction. class and tags can be sets.

add(param_set)
Param | ParamSet -> self Adds a single param or another set to this set, all params are cloned.

subset(ids)
character -> self Changes the current set to the set of passed IDs.

test(x), check(x), assert(x)
Three checkmate-like check-functions. Take a named list. A point x is feasible, if it configures a subset of params, all individual param constraints are satisfied and all dependencies are satisfied. Params for which dependencies are not satisfied should not be part of x.

add_dep(id, on, cond)
character(1), character(1), Condition -> self Adds a Dependency to this set, so that param id now depends on param on.

S3 methods and type converters

as.data.table()
Compact representation as datatable. Col types are:

id: character

lower, upper: double

values: list col, with NULL elements

special_vals: list col of list

is_bounded: logical

default: list col, with NULL elements

storage_type: character

tags: list col of character vectors